Add files generated by flint/bootstrap.sh
They are gitignored in flint's repo, but they should be provided
to end users compiling the library (like how they are included
in tarballs).

There is another file generated by bootstrap.sh: aclocal.m4.
But it's essentially a temporary files passed between different autotools steps.
It's not included in the tarball so we are not committing it either.